moved to v2, versioned tests and serializers as well
